tps-demo: widen getMultiplayerAuthority comparison to Long for iOS
iOS returns Int for get_multiplayer_authority (int32); compare via .toLong() so
it matches getUniqueId()?.toLong() on both platforms.

+++ tps-demo-kanama/kotlin-src/PlayerInputSynchronizer.kt
@@ -72,7 +72,7 @@ class PlayerInputSynchronizer(godotObject: MemorySegment) :
 
     @OnReady
     fun ready() {
-        if (self.getMultiplayerAuthority() == self.getMultiplayer()?.getUniqueId()?.toLong()) {
+        if (self.getMultiplayerAuthority().toLong() == self.getMultiplayer()?.getUniqueId()?.toLong()) {
             cameraCamera?.makeCurrent()
             if (net.multigesture.kanama.api.DisplayServer.getName() != "headless") {
                 Input.setMouseMode(Input.MOUSE_MODE_CAPTURED)
